Different searchers suggest same initial random configs. New methods in baselines
