+++ syne_tune/optimizer/schedulers/searchers/kde/kde_searcher.py
@@ -16,10 +16,19 @@ import numpy as np
 import statsmodels.api as sm
 import scipy.stats as sps
 
-from syne_tune.optimizer.schedulers.searchers import SearcherWithRandomSeed
+from syne_tune.optimizer.schedulers.searchers.searcher import (
+    SearcherWithRandomSeed,
+    sample_random_configuration,
+)
 import syne_tune.config_space as sp
 from syne_tune.optimizer.schedulers.searchers.bayesopt.utils.debug_log import (
     DebugLogPrinter,
+)
+from syne_tune.optimizer.schedulers.searchers.utils.hp_ranges_factory import (
+    make_hyperparameter_ranges,
+)
+from syne_tune.optimizer.schedulers.searchers.bayesopt.tuning_algorithms.common import (
+    ExclusionList,
 )
 
 logger = logging.getLogger(__name__)
@@ -55,7 +64,10 @@ class KernelDensityEstimator(SearcherWithRandomSeed):
     :param mode: Mode to use for the metric given, can be "min" or "max". Is
         obtained from scheduler in :meth:`configure_scheduler`. Defaults to "min"
     :param num_min_data_points: Minimum number of data points that we use to fit
-        the KDEs. If set to ``None``, we set this to the number of hyperparameters.
+        the KDEs. As long as less observations have been received in
+        :meth:`update`, randomly drawn configurations are returned in
+        :meth:`get_config`.
+        If set to ``None``, we set this to the number of hyperparameters.
         Defaults to ``None``.
     :param top_n_percent: Determines how many datapoints we use to fit the first
         KDE model for modeling the well performing configurations.
@@ -142,8 +154,13 @@ class KernelDensityEstimator(SearcherWithRandomSeed):
         self.num_min_data_points = (
             len(self.vartypes) if num_min_data_points is None else num_min_data_points
         )
-        assert self.num_min_data_points >= len(self.vartypes)
+        assert self.num_min_data_points >= len(
+            self.vartypes
+        ), f"num_min_data_points = {num_min_data_points}, must be >= {len(self.vartypes)}"
         self._resource_attr = kwargs.get("resource_attr")
+        # Used for sampling initial random configs, and to avoid duplicates
+        self._hp_ranges = make_hyperparameter_ranges(self.config_space)
+        self._excl_list = ExclusionList.empty_list(self._hp_ranges)
         # Debug log printing (switched on by default)
         debug_log = kwargs.get("debug_log", True)
         if isinstance(debug_log, bool):
@@ -253,6 +270,17 @@ class KernelDensityEstimator(SearcherWithRandomSeed):
             msg = f"Update for trial_id {trial_id}: metric = {metric_val:.3f}"
             logger.info(msg)
 
+    def _get_random_config(
+        self, exclusion_list: Optional[ExclusionList] = None
+    ) -> dict:
+        if exclusion_list is None:
+            exclusion_list = self._excl_list
+        return sample_random_configuration(
+            hp_ranges=self._hp_ranges,
+            random_state=self.random_state,
+            exclusion_list=exclusion_list,
+        )
+
     def get_config(self, **kwargs) -> Optional[dict]:
         suggestion = self._next_initial_config()
         if suggestion is None:
@@ -261,17 +289,15 @@ class KernelDensityEstimator(SearcherWithRandomSeed):
             if models is None or self.random_state.rand() < self.random_fraction:
                 # return random candidate because a) we don't have enough data points or
                 # b) we sample some fraction of all samples randomly
-                suggestion = {
-                    k: v.sample() if isinstance(v, sp.Domain) else v
-                    for k, v in self.config_space.items()
-                }
+                suggestion = self._get_random_config()
             else:
                 self.bad_kde = models[0]
                 self.good_kde = models[1]
                 l = self.good_kde.pdf
                 g = self.bad_kde.pdf
 
-                acquisition_function = lambda x: max(1e-32, g(x)) / max(l(x), 1e-32)
+                def acquisition_function(x):
+                    return max(1e-32, g(x)) / max(l(x), 1e-32)
 
                 current_best = None
                 val_current_best = None
@@ -322,12 +348,17 @@ class KernelDensityEstimator(SearcherWithRandomSeed):
                             "candidate has non finite acquisition function value"
                         )
 
-                    if val_current_best is None or val_current_best > val:
-                        current_best = candidate
+                    config = self._from_feature(candidate)
+                    if (
+                        val_current_best is None or val_current_best > val
+                    ) and not self._excl_list.contains(config):
+                        current_best = config
                         val_current_best = val
 
-                suggestion = self._from_feature(feature_vector=current_best)
+                suggestion = current_best
 
+        if suggestion is not None:
+            self._excl_list.add(suggestion)
         return suggestion
 
     def _train_kde(self, train_data, train_targets):
