+++ tst/blackbox_repository/test_data_consistency.py
@@ -21,7 +21,7 @@ from syne_tune.blackbox_repository.simulated_tabular_backend import (
 from syne_tune.blackbox_repository import load_blackbox, add_surrogate
 from syne_tune.blackbox_repository.utils import metrics_for_configuration
 from syne_tune.blackbox_repository.blackbox_tabular import BlackboxTabular
-from syne_tune.optimizer.schedulers.searchers.searcher import RandomSearcher
+from syne_tune.optimizer.schedulers.searchers import RandomSearcher
 
 
 @dataclass
