+++ tst/schedulers/test_random_searcher.py
@@ -10,7 +10,7 @@
 # on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
 # express or implied. See the License for the specific language governing
 # permissions and limitations under the License.
-from syne_tune.optimizer.schedulers.searchers.searcher import RandomSearcher
+from syne_tune.optimizer.schedulers.searchers import RandomSearcher
 from syne_tune.config_space import choice, randint
